+++ lib/src/protocol/language_server/interface.dart
@@ -8,6 +8,8 @@ abstract class LanguageServer {
   Future<Null> shutdown() async {}
   void exit() {}
 
+  Future<ServerCapabilities> initialize(int clientPid, String rootUri,
+      ClientCapabilities clientCapabilities, String trace);
   void textDocumentDidOpen(TextDocumentItem document) {}
   void textDocumentDidChange(VersionedTextDocumentIdentifier documentId,
       List<TextDocumentContentChangeEvent> changes) {}
@@ -22,7 +24,9 @@ abstract class LanguageServer {
       ReferenceContext context);
   Future<Hover> textDocumentHover(
       TextDocumentIdentifier documentId, Position position);
-  Future<List<Command>> textDocumentCodeAction(TextDocumentIdentifier documentId,
-      Range range, CodeActionContext context);
+  Future<List<Command>> textDocumentCodeAction(
+      TextDocumentIdentifier documentId,
+      Range range,
+      CodeActionContext context);
   Stream<Diagnostics> get diagnostics;
 }